requirements.txt
google-genai
fastapi
uvicorn[standard]
python-multipart
pandas
openpyxl
pillow
python-dotenv
PyMuPDF
google-generativeai